requirements.txt
absl-py==0.11.0
aliyun-python-sdk-core==2.13.30
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.14.0
appdirs==1.4.4
astor==0.8.1
async-generator==1.10
audioread==2.1.9
backcall==0.2.0
blessings==1.7
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
chumpy @ git+https://github.com/hassony2/chumpy.git@c0118792626e7e18f485b115983e3f87f287517a
click==7.1.2
codecov==2.1.11
coloredlogs==15.0
colour-runner==0.1.1
coverage==5.3.1
crcmod==1.7
cryptography==3.2.1
cycler==0.10.0
Cython==0.29.21
decorator==4.4.2
deepdiff==5.0.2
defusedxml==0.6.0
distlib==0.3.1
easydict==1.9
einops==0.3.0
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
fsspec==0.8.5
future==0.18.2
fvcore==0.1.2.post20201218
gast==0.4.0
gdown==3.12.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.28.1
grpcio-tools==1.28.1
h5py==3.1.0
hparams==0.3.0
humanfriendly==9.1
idna==2.10
importlib-metadata==2.1.1
iopath==0.1.2
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.35.0
Markdown==3.3.3
matplotlib==3.3.3
msgpack==1.0.2
numba==0.52.0
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opendr==0.78
ordered-set==4.0.2
oss2==2.13.1
pafy==0.5.5
Pillow==8.0.1
pluggy==0.13.1
pooch==1.3.0
portalocker==2.0.0
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycryptodome==3.9.9
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.1
pytorch-lightning==1.1.2
pytorch3d==0.3.0
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
retry==0.9.2
rootpath==0.1.1
rsa==4.6
scikit-learn==0.24.0
scipy==1.5.4
sklearn==0.0
smplx==0.1.26
SoundFile==0.10.3.post1
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==2.1
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.2
torchgeometry==0.1.2
torchvision==0.8.2
tox==3.20.1
tqdm==4.54.1
typeguard==2.10.0
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
watchdog==1.0.2
webencodings==0.5.1
Werkzeug==0.15.5
wrapt==1.12.1
yacs==0.1.8
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git